Organized code by creating new file.
- Created new file called `gpt_utils.py` which holds all OpenAI functions which were originally in opportunity.py

- Moved `format_opportunities()` to opportunity.py as it fits better there.

- Organization fixes Issue #8
